backend/pyproject.toml
[project]
name = "noctem-backend"
version = "0.1.0"
description = "Noctem Virtual Dispatcher — FastAPI backend"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "httpx>=0.28.0",
    "python-dotenv>=1.0.0",
    "supabase>=2.11.0",
    "pydantic>=2.10.0",
    "openai>=1.60.0",
]
